Rename `getFileName` to `getFileBasename` for clarity
