+++ packages/media/src/getFileBasename.js
@@ -21,7 +21,8 @@
  * @param {string} file.name File name.
  * @return {string} File name without extension.
  */
-const getFileName = ({ name = '' }) =>
-  name.includes('.') ? name.split('.').slice(0, -1).join('.') : name;
+function getFileBasename({ name = '' }) {
+  return name.includes('.') ? name.split('.').slice(0, -1).join('.') : name;
+}
 
-export default getFileName;
+export default getFileBasename;
